django-backend/pyproject.toml
[project]
name = "django-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.7",
    "django-environ>=0.12.0",
    "django-ninja>=1.3.0",
    "django-vite>=3.1.0",
    "django-extensions>=3.2.3",
    "pillow>=11.1.0",
    "pydantic[email]>=2.10.6",
    "django-cors-headers>=4.7.0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.2",
]

[tool.ruff]
line-length = 79
